requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0

# For CLIP (retrieval)
sentence-transformers>=2.2.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Optional: for downloading Stanford Cars
gdown>=4.7.0